+++ src/main/java/kr/hhplus/be/interfaces/event/PaymentEventListener.java
@@ -1,9 +1,8 @@
-package kr.hhplus.be.application.event;
+package kr.hhplus.be.interfaces.event;
 
-import kr.hhplus.be.domain.event.PaymentSuccessEvent;
+import kr.hhplus.be.infrastructure.event.PaymentSuccessEvent;
 import kr.hhplus.be.domain.order.client.PlatformClient;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
