+++ src/test/java/kr/hhplus/be/application/event/PaymentEventListenerTest.java
@@ -3,8 +3,9 @@ package kr.hhplus.be.application.event;
 import kr.hhplus.be.application.order.dto.OrderInfo;
 import kr.hhplus.be.application.order.dto.OrderItemInfo;
 import kr.hhplus.be.application.order.dto.OrderItems;
-import kr.hhplus.be.domain.event.PaymentSuccessEvent;
+import kr.hhplus.be.infrastructure.event.PaymentSuccessEvent;
 import kr.hhplus.be.domain.order.client.PlatformClient;
+import kr.hhplus.be.interfaces.event.PaymentEventListener;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
